+++ app/lib/shopify.server.js
@@ -1,23 +1,33 @@
 import "@shopify/shopify-app-remix/adapters/node";
 import {
+  DeliveryMethod,
   shopifyApp,
 } from "@shopify/shopify-app-remix/server";
-import { MongoDBSessionStorage } from "@shopify/shopify-app-session-storage-mongodb";
+import { PrismaSessionStorage } from "@shopify/shopify-app-session-storage-prisma";
+import prisma from "../lib/db.server"
 
 const shopify = shopifyApp({
   apiKey: process.env.SHOPIFY_API_KEY,
   apiSecretKey: process.env.SHOPIFY_API_SECRET || "",
   scopes: process.env.SHOPIFY_API_SCOPES?.split(","),
   appUrl: process.env.SHOPIFY_APP_URL,
-  sessionStorage: new MongoDBSessionStorage(process.env.MONGO_URL, "images"),
+  sessionStorage: new PrismaSessionStorage(prisma),
+  webhooks: {
+    APP_UNINSTALLED: {
+      deliveryMethod: DeliveryMethod.Http,
+      callbackUrl: "/webhooks",
+    },
+  },
   hooks: {
-    afterAuth: async ({session}) => {
-
+    afterAuth: async ({ session }) => {
+      shopify.registerWebhooks({ session });
     },
   },
 });
 
 export default shopify;
 
-export const login = shopify.login;
+
 export const authenticate = shopify.authenticate;
+export const login = shopify.login;
+export const registerWebhooks = shopify.registerWebhooks;