+++ app/routes/webhooks.jsx
@@ -1,0 +1,33 @@
+import { authenticate } from "../lib/shopify.server";
+import sessionService from "../services/sessionService";
+
+export const action = async ({ request }) => {
+  const { topic, shop, session, admin } = await authenticate.webhook(request);
+
+  if (!admin) {
+    // The admin context isn't returned if the webhook fired after a shop was uninstalled.
+    throw new Response();
+  }
+
+  switch (topic) {
+    case "APP_UNINSTALLED":
+      if (session) {
+        try {
+          console.log("APP_UNINSTALLED:");
+
+          await sessionService.deleteSession(shop);
+        } catch (error) {
+          console.log("error:", error);
+        } 
+      }
+
+      break;
+    case "CUSTOMERS_DATA_REQUEST":
+    case "CUSTOMERS_REDACT":
+    case "SHOP_REDACT":
+    default:
+      throw new Response("Unhandled webhook topic", { status: 404 });
+  }
+
+  throw new Response();
+};